#52 Rename submodule (api → core)
