+++ tests/test_core.py
