+++ typespecs/core.py
